Share compressed test fixture loader
